Complete GPU parity fixes, test coverage, and cleanup
GPU zero guards: Add powi(0) n==0 guard to all 5 CUDA kernel modes
and all 4 WGSL shaders. Add Powf a=0 da_eps guard to CUDA
tangent_reverse. Add Hypot r=0 guards to CUDA/WGSL tangent kernels.

GPU CPU parity: Fix _sign/signum to return ±1 (not 0) at x=0 and
propagate NaN, matching Rust's f64::signum. Fix ABS derivative to
match across all shader modes. Fix asinh in tangent shaders to use
stable |x| formula.

Remaining fixes: checkpoint beta overflow via checked_mul, 3 more u32
buffer overflow sites in wgpu_backend, biharmonic reverted to diagonal
4th-order with accurate docstring (mixed-partial jet evaluation not
yet supported).

Test coverage: Add Reverse-mode powi(0)/max/min NaN tests, Rem db
partial test, tighten atan2 assertion to check exact derivative value.

+++ src/gpu/wgpu_backend.rs
@@ -1056,7 +1056,7 @@ impl GpuBackend for WgpuContext {
             mapped_at_creation: false,
         });
 
-        let out_size = (batch * tape.num_outputs) as u64 * 4;
+        let out_size = (batch as u64) * (tape.num_outputs as u64) * 4;
         let tangent_out_buf = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("tangent_outputs"),
             size: out_size,
@@ -1215,8 +1215,8 @@ impl GpuBackend for WgpuContext {
                 usage: wgpu::BufferUsages::STORAGE,
             });
 
-        let buf_size = (batch_size * nv) as u64 * 4;
-        let grad_size = (batch_size * ni) as u64 * 4;
+        let buf_size = (batch_size as u64) * (nv as u64) * 4;
+        let grad_size = (batch_size as u64) * (ni as u64) * 4;
 
         let primals_work = self.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("pw"),
